Refactored tests for ValidateConnector
